requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Django==1.11
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
idna==2.8
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
